Namespace function parsehocon()
