+++ lib/puppet/functions/parsehocon.rb
@@ -1,32 +1,14 @@
 # frozen_string_literal: true
 
-# @summary
-#   This function accepts HOCON as a string and converts it into the correct
-#   Puppet structure
-#
-# @example How to parse hocon
-#   $data = parsehocon("{any valid hocon: string}")
-#
+# THIS FILE WAS GENERATED BY `rake regenerate_unamespaced_shims`
+
+# @summary DEPRECATED.  Use the namespaced function [`stdlib::parsehocon`](#stdlibparsehocon) instead.
 Puppet::Functions.create_function(:parsehocon) do
-  # @param hocon_string A valid HOCON string
-  # @param default An optional default to return if parsing hocon_string fails
-  # @return [Data]
-  dispatch :parsehocon do
-    param          'String', :hocon_string
-    optional_param 'Any',    :default
+  dispatch :deprecation_gen do
+    repeated_param 'Any', :args
   end
-
-  def parsehocon(hocon_string, default = :no_default_provided)
-    require 'hocon/config_factory'
-
-    begin
-      data = Hocon::ConfigFactory.parse_string(hocon_string)
-      data.resolve.root.unwrapped
-    rescue Hocon::ConfigError::ConfigParseError => e
-      Puppet.debug("Parsing hocon failed with error: #{e.message}")
-      raise e if default == :no_default_provided
-
-      default
-    end
+  def deprecation_gen(*args)
+    call_function('deprecation', 'parsehocon', 'This function is deprecated, please use stdlib::parsehocon instead.')
+    call_function('stdlib::parsehocon', *args)
   end
 end
