+++ spec/functions/parsehocon_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-describe 'parsehocon' do
+describe 'stdlib::parsehocon' do
   it { is_expected.to run.with_params('').and_return({}) }
   it { is_expected.to run.with_params('valid hocon: string').and_return('valid hocon' => 'string') }
   it { is_expected.to run.with_params('invalid').and_raise_error(Hocon::ConfigError::ConfigParseError) }
